refactor(api): split event and application API calls into separate files and update imports
