+++ pages/apply.tsx
@@ -11,7 +11,7 @@ import {
   getCurrentTerm,
   getCurrentUserApplication,
   createOrUpdateApplication,
-} from "@/utils/apiCalls";
+} from "@/utils/applicationApiCalls";
 
 // Form Components
 import AppIntro from "@/components/forms/application/AppIntro";
